+++ frontend/javascripts/viewer/model/sagas/volume/helpers.ts
@@ -10,12 +10,11 @@ import Constants, {
   type Vector3,
 } from "viewer/constants";
 import { mayEditAnnotation } from "viewer/model/accessors/annotation_accessor";
+import { getLayerByName, getMagInfo } from "viewer/model/accessors/dataset_accessor";
 import {
-  getLayerByName,
-  getMagInfo,
   getTransformedDatasetBoundingBox,
-} from "viewer/model/accessors/dataset_accessor";
-import { getTransformsForLayer } from "viewer/model/accessors/dataset_layer_transformation_accessor";
+  getTransformsForLayer,
+} from "viewer/model/accessors/dataset_layer_transformation_accessor";
 import { enforceActiveVolumeTracing } from "viewer/model/accessors/volumetracing_accessor";
 import BoundingBox from "viewer/model/bucket_data_handling/bounding_box";
 import type DataCube from "viewer/model/bucket_data_handling/data_cube";
